+++ src/Pim/MultiVendorProduct/MultiVendorProduct.php
@@ -11,7 +11,6 @@ use Symfony\Component\Validator\ConstraintViolationInterface;
 use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Validation;
-use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use function theodorejb\polycast\to_int;
 use function theodorejb\polycast\to_string;