+++ src/Pim/MultiVendorProduct/MultiVendorProductFile.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * @copyright Copyright (c) Wizacha
+ * @license Proprietary
+ */
+
+namespace Wizaplace\SDK\Pim\MultiVendorProduct;
+
+use Wizaplace\SDK\Organisation\OrganisationFile;
+
+class MultiVendorProductFile extends OrganisationFile
+{
+}